// Views/src/components/financial/components/pix-expenses/components/list/dto/list-pix-expense-dto.ts
import { ItemsInterface } from "../../../../../../../shared/components/list-g/list/interfaces/items-interface";

export class ListPixExpenseDto {
  id: ItemsInterface;
  expenseDay: ItemsInterface;
  expenseDayToFilter: ItemsInterface;
  price: ItemsInterface;
  pixOutId: ItemsInterface;
  benefitedName: ItemsInterface;
}